% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_matching.R
\name{assn}
\alias{assn}
\title{Assignment for a pipepline}
\usage{
assn(value, x, pos = 1, ...)
}
\arguments{
\item{value}{to assign}

\item{x}{name}

\item{pos}{position in search() list in which to assign (default = 1)}

\item{...}{other arguments to 'assign'}
}
\description{
Assignment for a pipepline
}
