% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_chat.R
\name{chat2xl}
\alias{chat2xl}
\title{Create excel file from Zoom chat file}
\usage{
chat2xl(file, outfile = NULL, overwrite = FALSE)
}
\arguments{
\item{file}{chat file}

\item{outfile}{default 'file' with '.txt' suffix replaced with .xlsx}
}
\description{
Excel file allows sorting by poster so all submissions 
are contiguous and in time order
}
