% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_function.R
\name{coding_function}
\alias{coding_function}
\title{General coding matrix for factors}
\usage{
coding_function(x, levs = levels(as.factor(x)), intercept = rbind(c(1,
  rep(0, length(levs) - 1))), constraints = NULL,
  estimates = diag(length(levs))[-1, , drop = FALSE],
  tolerance = 1e-16, debug = FALSE)
}
\description{
This function creates functions (closures) that implement a flexible coding
matrix for a factor.
}
\details{
The coding function can be designed with specified constraints and 
specified estimated model parameters.

The coding function can also be used to generate portions of hypothesis 
matrices to estimate interpretable functions of parameters.
}
