% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{combine_small_categories}
\alias{combine_small_categories}
\alias{gr}
\alias{combine_later_categories}
\title{Combine small categories}
\usage{
combine_small_categories(x, nlevels = 3, other = "Other", verbose = FALSE, ...)

combine_later_categories(x, nlevels = 3, other = "Other", verbose = FALSE, ...)
}
\arguments{
\item{x}{factor or character to be condensed}

\item{nlevels}{number of levels to keep}

\item{other}{name for 'other' category}

\item{verbose}{default FALSE}
}
\description{
Combine small categories
}
\section{Functions}{
\itemize{
\item \code{combine_later_categories}: later categories
}}

