% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{compare_packages}
\alias{compare_packages}
\title{List functions in all of two or more packages}
\usage{
compare_packages(places)
}
\arguments{
\item{sources}{a character vector or list with characters and/or numeric values
identifying two or more sources in which to look for common functions as
package names in quotes. "" refers to the Global environment}
}
\value{
The definitions of functions in two or more sources
}
\description{
Useful to check discrepancies among objects with the same name in different
packages. Use \link{ga} to
}
\examples{
compare_packages(c('spida2','gnew'))  
}
