% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{droplists}
\alias{droplists}
\title{Remove lists from a data frame}
\usage{
droplists(x, ...)
}
\arguments{
\item{x}{a data frame}

\item{quiet}{do not warn if TRUE, default FALSE}
}
\value{
data set dropping variables that are lists
}
\description{
Many methods for data frames fail is a variable is a list but lists can be very
useful for many purposes.
}
