% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{dropone}
\alias{dropone}
\title{Drop one case or cluster}
\usage{
dropone(
  fit,
  form = NULL,
  FUN = if (inherits(fit, "lme")) fixef else coef,
  data = getData(fit),
  ...
)
}
\arguments{
\item{fit}{a fitted model}

\item{form}{a formula to}
}
\description{
Uses parallel::mclapply which is not effective on PCs
}
