% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan.R
\name{getAll}
\alias{getAll}
\title{Get samples and sampler parameters from a stanfit object}
\usage{
getAll(sf)
}
\arguments{
\item{sf}{a stanfit object}
}
\description{
Get samples and sampler parameters from a stanfit object
}
