% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnew.R
\docType{package}
\name{gnew-package}
\alias{gnew}
\alias{gnew-package}
\title{New candidate functions for spida2}
\description{
Currently working on \pkg{gnew} to add
functions \code{\link{getG}}, etc.
}
\author{
\strong{Maintainer}: Georges Monette \email{georges.monette@gmail.com}

}
