% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{grepf}
\alias{grepf}
\title{grep patterns and replace with name}
\usage{
grepf(matches, pats)
}
\arguments{
\item{matches}{object in which to find regular expression matches}

\item{pattern}{to match}
}
\description{
grep patterns and replace with name
}
