% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{kbl_array}
\alias{kbl_array}
\title{Use kableExtra::kbl to print an array.}
\usage{
kbl_array(a, MARGIN, ...)
}
\arguments{
\item{a}{array}

\item{MARGIN}{a vector of dimension along which the array will be collapse}
}
\description{
Requires version 1.2.0 or later of kableExtra
from devtools::install_github("haozhu233/kableExtra")
}
\details{
See https://haozhu233.github.io/kableExtra/awesome_table_in_pdf.pdf. 
The output can be postprocessed with other function in kableExtra.
}
\examples{
\dontrun{
kbl_array(Titanic, 3:4, longtable = T)
}
}
