% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{list_array}
\alias{list_array}
\title{Turn an array into a list of smaller arrays sectioning along a margin}
\usage{
list_array(a, MARGIN)
}
\arguments{
\item{a}{array}

\item{MARGIN}{vector of dimensions along which to section}
}
\value{
list of subarrays with names consisting of headings for each subarray
}
\description{
Turn an array into a list of smaller arrays sectioning along a margin
}
\examples{
a <- Titanic
library(spida2)
tab(a)
tab(aperm(a,c(2,4,1,3)))
tab(aperm(a,c(2,4,1,3)), pct = c(1,3,4))
round(tab(aperm(a,c(2,4,1,3)), pct = c(1,3,4)),1)
list_array(round(tab(aperm(a,c(2,4,1,3)), pct = c(1,3,4)),1),3:4)

list_array(tab(a),3:4)
list_array(a,3:4)
list_array(tab(a, pct = c(1,3:4),3:4)
}
