% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{order_by_freq}
\alias{order_by_freq}
\title{Order a factor using the frequency of levels}
\usage{
order_by_freq(x, order = is.ordered(x))
}
\arguments{
\item{x}{a vector: numerical, character, factor, etc.}
}
\value{
a factor with levels ordered by frequency (largest first)
}
\description{
Order a factor using the frequency of levels
}
\examples{
  zf <- letters[c(1,1,2,2,2,3,2,4,5,6)]
zf
table(zf)
zo <- order_by_freq(zf)
zo
table(zo)
attributes(zo)
}
