% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_matching.R
\name{overlap_distance}
\alias{overlap_distance}
\title{Distance between two strings using weighted overlap of words}
\usage{
overlap_distance(strings1, strings2, split = " ")
}
\arguments{
\item{a}{set of strings from first corpus}

\item{b}{set of strings from second corpus}

\item{weights}{to be given to 'intersection' and 'union' matching
distance respectively. If NULL (default), both distances are
returned.}
}
\description{
Distance between two strings using weighted overlap of words
}
