% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{pr}
\alias{pr}
\alias{pr.default}
\alias{pr.array}
\alias{pr.table}
\title{Alternative printing for special purposes}
\usage{
pr(x, ...)

\method{pr}{default}(x, ...)

\method{pr}{array}(x, ...)

\method{pr}{table}(x, ...)
}
\arguments{
\item{x}{an object to be printed}

\item{...}{additional arguments}
}
\description{
Alternative printing for special purposes
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method

\item \code{array}: array method

\item \code{table}: table method
}}

\examples{
pr(Titanic)
}
