% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{print.summ.lm}
\alias{print.summ.lm}
\title{Adaptation of print.summary.lm}
\usage{
\method{print}{summ.lm}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\description{
Prints VIF from car instead of correlation matrix
}
