% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_matching.R
\name{printup}
\alias{printup}
\title{Print a nested data frame}
\usage{
printup(
  dd,
  form,
  sortform = NULL,
  row.names = FALSE,
  right = FALSE,
  sep = "======================================",
  ...
)
}
\arguments{
\item{dd}{a nested data frame}

\item{form}{a formula for grouping variables evaluated in dd}

\item{sortform}{an optional formal to sort rows within subgroups}

\item{sep}{separator between clusters}
}
\description{
Print a nested data frame
}
