% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{render_this}
\alias{render_this}
\title{Render with YAML is separate file}
\usage{
render_this(clean = TRUE)
}
\arguments{
\item{clean}{passed to \code{\link[rmarkdown]{render}}}
}
\description{
Should be used from a commented line in a file in RStudio. e.g.
}
\details{
\code{# gnew::render_this()}

Pre catenates the contents of YAML.R to the file as
filename_out.R and render using R markdown.

Uses spida2::here to identify the names of the file from
which it was called
}
