% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{slice}
\alias{slice}
\title{Take a slice of an array selecting indices along a margin}
\usage{
slice(a, MARGIN, ind)
}
\arguments{
\item{a}{an array}

\item{MARGIN}{a single dimension along which to select a slice (i.e. a section)}

\item{ind}{the level at which to select the slice. If ind is out of range,
slice returns NAs with the correct structure}
}
\description{
Performs the same role as selecting with subscripts except that
this is easier to use when dimension of the array is not known.
Its application is limited, however, to selecting single slices
along a single dimension.
}
\examples{
Titanic
slice(Titanic, 1, 1)  
slice(slice(Titanic, 1, 1), 1, 2)  
}
