% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{tablemissing}
\alias{tablemissing}
\title{Patterns of missing values by variable}
\usage{
tablemissing(
  x,
  sortby = "variable",
  bot = 6,
  mar = c(bot, 1, 1.5, 1),
  adj = c(0, -0.1),
  xpd = TRUE,
  srt = -60,
  cex = 0.8,
  main = "Missing Value Patterns",
  ...
)
}
\arguments{
\item{data}{frame to analyze}
}
\description{
Patterns of missing values by variable
}
