% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{tonum}
\alias{tonum}
\alias{c2f}
\alias{f2c}
\title{Turn character or factor to numeric}
\usage{
tonum(x)

c2f(x, ...)

f2c(x)
}
\arguments{
\item{x}{object to be turned into a character if 'is.factor' is true}

\item{...}{arguments to factor}
}
\value{
a numeric value represented by 'x', or NA

factor(x)

factor(x)
}
\description{
Turn character or factor to numeric

Character to factor

Character to factor
}
