% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coding_function.R
\name{weighted_coding}
\alias{weighted_coding}
\title{Weighted coding matrix}
\usage{
weighted_coding(x)
}
\arguments{
\item{x}{a factor used as the corpus}
}
\description{
Produce factor coding function so that the intercept is a weighted sum
of factor means using weights proportional to the frequency of 
factor levels in a corpus independent of the data. This has the
advantage of working consistently on different data sets.
}
