% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{yearsdiff}
\alias{yearsdiff}
\title{Difference of two dates in years}
\usage{
yearsdiff(date1, date2, ...)
}
\arguments{
\item{date1, date2}{Date objects or objects that can be coerced to Dates}
}
\value{
difference: date2 - date 1 in years
}
\description{
date2 - date1 in years
}
