% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_chat.R
\name{chat2xl}
\alias{chat2xl}
\title{Create excel file from Zoom chat file}
\usage{
chat2xl(file, outfile = NULL, overwrite = FALSE)
}
\arguments{
\item{file}{chat file, usually "chat.txt" or "meeting_saved_chat.txt"}

\item{outfile}{default 'file' with '.txt' suffix replaced with .xlsx}

\item{overwrite}{will an existing ".xlsx" file be overwritten, default FALSE}
}
\description{
Reads a Zoom "chat.txt" file (recorded portion of meeting)
or a "meeting_saved_chat.txt" file (entire meeting)
and writes a '.xlsx' Excel file in which lines are sorted by
participant and by time.
}
\details{
When a participant submits a chat message in multiple lines, the subsequent
lines may not be identified with the name of the participant, or they
may be separated by other chat submissions from other participants.
\code{chat2xl} resolves these problems and creates a file in which all
submissions are labelled with the identifier of the participant and
and ordered by participant and by time. When viewed, the excel file can be reordered
by time if desired.
}
