% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_matching.R
\name{rnd}
\alias{rnd}
\title{Round numeric values in a a data frame}
\usage{
rnd(x, ...)
}
\arguments{
\item{x}{data frame}

\item{...}{other arguments to round if x has mode double}
}
\description{
Round numeric values in a a data frame
}
