% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{tonum}
\alias{tonum}
\alias{c2f}
\alias{f2c}
\alias{combine_small_categories}
\alias{gr}
\title{Turn character or facator to numeric}
\usage{
tonum(x)

c2f(x, ...)

f2c(x)

combine_small_categories(x, nlevels = 3, other = "Other", verbose = FALSE, ...)
}
\arguments{
\item{x}{factor or character to be condensed}

\item{...}{arguments to factor}

\item{nlevels}{number of levels to keep}

\item{other}{name for 'other' category}

\item{verbose}{default FALSE}
}
\value{
a numeric value represented by 'x', or NA

factor(x)

factor(x)
}
\description{
Turn character or facator to numeric

Character to factor

Character to factor

Combine small categories
}
