% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{tonum}
\alias{tonum}
\alias{c2f}
\alias{Turn}
\alias{character}
\alias{or}
\alias{facator}
\alias{to}
\alias{numeric}
\alias{f2c}
\alias{combine_small_categories}
\title{Character to factor}
\usage{
tonum(x)

c2f(x, ...)

f2c(x)

combine_small_categories(x, nlevels = 3, other = "Other",
  verbose = FALSE, ...)
}
\arguments{
\item{x}{character or factor representing a numeric value}

\item{...}{arguments to factor}

\item{nlevels}{number of levels to keep}

\item{other}{name for 'other' category}

\item{verbose}{default FALSE}

\item{x}{object to be turned into a factor if 'is.character' is true}

\item{x}{object to be turned into a character if 'is.factor' is true}

\item{x}{factor or character to be condensed}
}
\value{
a numeric value represented by 'x', or NA

factor(x)

factor(x)
}
\description{
Character to factor

Character to factor

Combine small categories
}
